new web services

+++ Analyser/StaticCodeAnalyser/StaticCodeAnalysisReportsCSVMerger/StaticCodeAnalysisReportsCSVMerger.cs
@@ -11,22 +11,26 @@ namespace StaticCodeAnalysisReportsCSVMergerLib
 {
     public class StaticCodeAnalysisReportsCSVMerger
     {
-        public void WriteReportsToCSV(IStaticCodeAnalysisToolParser parser, string report, string outfile)
+        public bool WriteReportsToCSV(IStaticCodeAnalysisToolParser parser, string report, string outfile)
         {
-                string[] lines = File.ReadAllLines(report);
-                lines = parser.ParseReportToCSV(lines);
+            if (!File.Exists(report))
+                return false;
+            string[] lines = File.ReadAllLines(report);
+            lines = parser.ParseReportToCSV(lines);
 
-                FileStream f = new FileStream(outfile, FileMode.OpenOrCreate);
-                StreamWriter s = new StreamWriter(f);
-                f.Seek(f.Length, SeekOrigin.Begin);
+            FileStream f = new FileStream(outfile, FileMode.OpenOrCreate);
+            StreamWriter s = new StreamWriter(f);
+            f.Seek(f.Length, SeekOrigin.Begin);
 
-                foreach (string line in lines)
-                {
-                    s.WriteLine(line);
-                }
+            foreach (string line in lines)
+            {
+                s.WriteLine(line);
+            }
 
-                s.Close();
-                f.Close();
+            s.Close();
+            f.Close();
+
+            return true;
         }
     }
 }
